requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.35
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
Jinja2==3.1.4
qiniu==7.12.0
pydantic==2.9.2
email-validator==2.2.0